refactor(645W8H): extract shared re-entry lib (deferred from scenario 1.1)
Pure functions live in .safeword/hooks/lib/re-entry.ts so both
session-start-reentry.ts (Slice 2) and the upcoming statusline/reentry.ts
GREEN for scenario 8.2 can share them without duplication:

- parseLogLine + Entry type
- listOtherSessionTranscripts
- readRecentEditedFiles
- getDirtyFiles
- normalizeRelative
- detectConflictFiles

session-start-reentry.ts now imports from ./lib/re-entry; rendering
(renderBrief, renderConflictWarning) stays inline since it's
session-start-specific. All 19 re-entry tests still green. Parity 106
pairs + 3 contracts; schema test 25/25.
