+++ .safeword/hooks/session-start-reentry.ts
@@ -16,9 +16,10 @@
  * lightly around those files.
  */
 
-import { execSync } from 'node:child_process';
-import { existsSync, readdirSync, readFileSync } from 'node:fs';
-import { basename, dirname, join, relative } from 'node:path';
+import { existsSync, readFileSync } from 'node:fs';
+import { join } from 'node:path';
+
+import { detectConflictFiles, type Entry, parseLogLine } from './lib/re-entry';
 
 interface HookInput {
   session_id?: string;
@@ -27,121 +28,12 @@ interface HookInput {
   transcript_path?: string;
 }
 
-interface Entry {
-  timestamp: string;
-  sessionId: string;
-  ticket: string;
-  nextImperative: string;
-}
-
-interface ToolUse {
-  type: string;
-  name?: string;
-  input?: { file_path?: string };
-}
-
-interface TranscriptEntry {
-  type?: string;
-  message?: { role?: string; content?: ToolUse[] };
-}
-
-// Canonical log line: `<ISO-ts> <session-id> ticket=<id>/<phase> Next: <imperative>`
-const LINE_REGEX = /^(\S+)\s+(\S+)\s+(ticket=\S+)\s+Next:\s+(.+)$/;
-const EDIT_TOOL_NAMES = new Set(['Edit', 'Write', 'MultiEdit']);
-const RECENT_TURNS = 10;
-
-function parseLogLine(line: string): Entry | null {
-  const match = LINE_REGEX.exec(line.trim());
-  if (!match) return null;
-  return {
-    timestamp: match[1],
-    sessionId: match[2],
-    ticket: match[3],
-    nextImperative: match[4],
-  };
-}
-
 function renderBrief(entries: Entry[], options: { fromAnotherSession?: boolean } = {}): string {
   const header = options.fromAnotherSession
     ? 'Re-entry brief — most recent entry (from another session):'
     : `Re-entry brief — last ${entries.length} entries from this session:`;
   const lines = entries.map(e => `- ${e.timestamp} [${e.ticket}] ${e.nextImperative}`);
   return `${header}\n${lines.join('\n')}`;
-}
-
-function listOtherSessionTranscripts(currentTranscriptPath: string): string[] {
-  const directory = dirname(currentTranscriptPath);
-  const currentBase = basename(currentTranscriptPath);
-  try {
-    return readdirSync(directory)
-      .filter(name => name.endsWith('.jsonl') && name !== currentBase)
-      .map(name => join(directory, name));
-  } catch {
-    return [];
-  }
-}
-
-function readRecentEditedFiles(transcriptPath: string): string[] {
-  const edited = new Set<string>();
-  try {
-    const raw = readFileSync(transcriptPath, 'utf8').trim();
-    if (raw.length === 0) return [];
-    const lines = raw.split('\n');
-    // Scan the last RECENT_TURNS * 5 lines as a coarse upper bound — one assistant
-    // turn often emits multiple tool-use events.
-    for (const line of lines.slice(-RECENT_TURNS * 5)) {
-      try {
-        const entry = JSON.parse(line) as TranscriptEntry;
-        if (entry.type !== 'assistant' || !entry.message?.content) continue;
-        for (const item of entry.message.content) {
-          if (item.type === 'tool_use' && item.name && EDIT_TOOL_NAMES.has(item.name)) {
-            const filePath = item.input?.file_path;
-            if (filePath) edited.add(filePath);
-          }
-        }
-      } catch {
-        // Skip malformed lines silently.
-      }
-    }
-  } catch {
-    // Transcript unreadable — no edits to report.
-  }
-  return [...edited];
-}
-
-function getDirtyFiles(cwd: string): string[] {
-  try {
-    const output = execSync('git status --porcelain', { cwd, encoding: 'utf8' });
-    return output
-      .split('\n')
-      .map(line => line.slice(3).trim())
-      .filter(line => line.length > 0);
-  } catch {
-    return [];
-  }
-}
-
-function normalizeRelative(filePath: string, cwd: string): string {
-  // Edit tool_use events typically use absolute paths; git status emits paths
-  // relative to the repo root (== cwd here).
-  if (filePath.startsWith(cwd)) {
-    return relative(cwd, filePath);
-  }
-  return filePath;
-}
-
-function detectConflictFiles(cwd: string, transcriptPath: string | undefined): string[] {
-  if (!transcriptPath) return [];
-  const dirtyFiles = new Set(getDirtyFiles(cwd));
-  if (dirtyFiles.size === 0) return [];
-  const overlap = new Set<string>();
-  for (const otherPath of listOtherSessionTranscripts(transcriptPath)) {
-    for (const editedFile of readRecentEditedFiles(otherPath)) {
-      const relativePath = normalizeRelative(editedFile, cwd);
-      if (dirtyFiles.has(relativePath)) overlap.add(relativePath);
-    }
-  }
-  return [...overlap];
 }
 
 function renderConflictWarning(files: string[]): string {
